pyproject.toml
[project]
name = "invoice-infer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "flask-cors>=5.0.1",
    "flask-sqlalchemy>=3.1.1",
    "pytz>=2025.1",
    "gevent>=24.11.1",
    "flask-restx>=1.3.0",
    "pdf2image>=1.17.0",
    "pydantic-ai>=0.0.40",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.11.0",
]

[tool.ruff]
exclude = [
    ".ruff_cache",
    ".git",
    ".ipynb_checkpoints",
    ".venv",
    ".vscode",
    "notebooks"
]
extend-include = []
line-length = 120
indent-width = 4
target-version = "py39"

[tool.ruff.lint]
ignore = ["ANN204", "ANN401", "E731", "D", "DTZ005", "BLE001","B008", "CPY001","COM812", "ERA001", "EM101","EM102", "FA","FBT", "G004", "UP", "TRY", "PTH123","ISC001" ]
select = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
